# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText - OpenAPI 3.0
  description: |-
    This is the OpenAPI specification for the WASAText web-app.
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /conversations/{conversationId}/messages:
    parameters:
      - name: conversationId
        in: path
        description: |-
            The id of the conversation
        schema:
          type: integer
        required: true
    post:
      tags: ["messages"]
      operationId: sendMessage
      summary: The user sends a message
      description: |-
        A new message is created in the specified conversation.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageText"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MessageImage"
        required: true
      responses:
        "201":
          description: |-
              Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageCreated"
        "400":
          description: |-
              Bad request - Invalid message content
        "401":
          description: |-
              Unauthorized
        "403":
          description: |-
              Forbidden - Not a participant in this conversation
        "404":
          description: Conversation not found
...
  /users:
    get:
      tags: ["users"]
      operationId: searchUsers
      summary: Search for users by username
      description: |
        Returns a list of users whose username matches the search query.
        Used to find users to start conversations with.
      parameters:
        - name: randomUser
          in: query
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 100
            example: "randomUser"
            pattern:  '^[A-Za-z0-9]{3,100}$'
          description: |-
              Search query (partial username match)
      responses:
        "200":
          description: List of matching users
          content:
            application/json:
              schema:
                description: |-
                    Schema of the 200 response to the "searchUsers" operation.
                type: array
                minItems: 0
                maxItems: 1000000000
                items:
                  description: |-
                      Each item of the array represents a user/outcome of the search
                  type: object
                  properties:
                    username:
                      description: |-
                          The username of a single object of the array.
                      type: string
                      example: "Maria"
                    photoUrl:
                      description: |-
                          The photoUrl of a single object of the array
                      type: string
                      format: url
                      example: "https://example.com/photos/user123.jpg"
        "401":
          description: Unauthorized
  /users/me/photo:
    put:
      tags: ["user"]
      operationId: setMyPhoto
      summary: Update my profile photo
      description: |-
        Updates the authenticated user's profile photo.
        The user is identified by the Bearer token.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UserPhoto"
      responses:
        "200":
          description: Photo updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPhotoUpdated"
        "400":
          description: Bad request - Invalid file format or size
          content:
            application/json:
              schema:
                description: |-
                    The 400 response to the "setMyPhoto" operation.
                type: object
                properties:
                  error:
                    description: |-
                        A string describing the error.
                    type: string
                    example: "Invalid image format. Supported: JPEG, PNG"
        "401":
          description: Unauthorized
        "413":
          description: Payload too large - Photo file too big
          content:
            application/json:
              schema:
                description: |-
                    The 413 schema of the response to the "setMyPhoto" operation.
                type: object
                properties:
                  error:
                    description: |-
                        A string describing the error.
                    type: string
                    example: "File size exceeds maximum allowed (5MB)"
  #GROUP MANAGEMENT
...
  /groups/{groupId}/name:
    put:
      tags: ["groups"]
      operationId: setGroupName
      summary: Update group name
      description: |-
        Updates the name of the specified group.
        Only group members can update the group name.
        The user is identified by the Bearer token.
      parameters:
        - name: groupId
          in: path
          description: The id of the group
          required: true
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupName"
      responses:
        "200":
          description: Group name updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupNameUpdated"
        "400":
          description: Bad request - Invalid group name
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Not a member of this group
          content:
            application/json:
              schema:
                description: |-
                    This schema represents the 403 response of the "setGroupName" operation.
                type: object
                properties:
                  error:
                    description: |-
                        This property represents a string describing the error.
                    type: string
                    example: "You must be a group member to change the name"
        "404":
          description: Group not found
  /groups/{groupId}/photo:
    put:
      tags: ["groups"]
      operationId: setGroupPhoto
      summary: Update group photo
      description: |-
        Updates the photo of the specified group.
        Only group members can update the group photo.
        The user is identified by the Bearer token.
      parameters:
        - name: groupId
          in: path
          description: The id of the group
          required: true
          schema:
            type: integer
            minimum: 1
            example: 42
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/GroupPhoto"
      responses:
        "200":
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupPhotoUpdated"
        "400":
          description: Bad request - Invalid file format or size
          content:
            application/json:
              schema:
                description: |-
                    This schema represents the 400 error of the "setGroupPhoto" operation.
                type: object
                properties:
                  error:
                    description: |-
                        A string describing the error.
                    type: string
                    example: "Invalid image format. Supported: JPEG, PNG"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - Not a member of this group
          content:
            application/json:
              schema:
                description: |-
                    This schema represents the 403 error of the "setGroupPhoto" operation.
                type: object
                properties:
                  error:
                    description: |-
                        A string describing the error.
                    type: string
                    example: "You must be a group member to change the photo"
        "404":
          description: Group not found
        "413":
          description: Payload too large - Photo file too big
          content:
            application/json:
              schema:
                description: |-
                    This schema represents the 413 error of the "setGroupPhoto" operation.
                type: object
                properties:
                  error:
                    description: |-
                        A string describing the error.
                    type: string
                    example: "File size exceeds maximum allowed (5MB)"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |-
        User identifier returned from login.
        Use the identifier as the bearer token.
  schemas:
...
    MessageText:
      description: |-
          This schema represents the text of a message, if it is a string.
      type: object 
      properties:
        content:
          description: |-
              The content of the message
          type: string
          example: "Ciao, come stai?"
          minLength: 1
          maxLength: 100000
          pattern: '^(.|[\r\n]){1,100000}$'
        replyToMessageId:
          description: |-
              The id of the message to which the message is a response (optional).
          type: integer
          format: int64
          minimum: 1
          example: 18284
      required: [content]
    MessageImage:
      description: |-
          This schema represents the image of a message.
      type: object
      properties:
        photo:
          description: |-
              This property represents the photo of a message.
          type: string
          format: binary
          minLength: 1
          maxLength: 100000000
      required: [photo]
    MessageCreated:
      description: |-
          This schema represents a message just created.
      type: object 
      properties:
        id:
          description: |-
              The id of the message.
          type: integer
          minimum: 1
          example: 18284
        checkmarks:
          description: |-
              The checkmarks of the message.
          type: integer
          maximum: 2
          example: 0
        timestamp:
          description: |-
              The timestamp of the message (when it has been sent).
          type: string
          format: date-time
        content:
          description: |-
              The content of the message (if it is a string).
          type: string
          example: "Ciao, come stai?"
        photoUrl:
          description: |-
              The content of the message (if it is a pic).
          type: string
          format: url
...
    UserPhoto:
      description: |-
          This schema represents the photo of a user. Hence it is characterized by a single propery: the photo itself.
      type: object
      required:
        - photo
      properties:
        photo:
          type: string
          format: binary
          description: |-
              Profile photo file (JPEG, PNG, etc.)
          minLength: 1
          maxLength: 100000000
    UserPhotoUpdated:
      description: |-
          This schema represents the photo of a user's profile. Hence it is characterized by a unique propery: the photoUrl itself.
      type: object
      properties:
        photoUrl:
          description: |-
              The url of the photo.
          type: string
          format: url
          example: "https://example.com/photos/user123.jpg"
    GroupName:
      description: |-
          This schema represents the name of a group. Hence it is characterized by a single property (the name itself).
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Gruppo Università 2025"
          minLength: 3
          maxLength: 50
          description: |-
              New group name
          pattern: "^[A-Za-z0-9 àèìòù]{3,50}$"
    GroupNameUpdated:
      description: |-
          This schema represents the new name of a particular group. Hence it is characterized by 2 properties: the group Id and the new name of the group.
      type: object
      properties:
        groupId:
          description: |-
              The id of the group.
          type: integer
          example: 42
        name:
          description: |-
              The new name of the group.
          type: string
          minLength: 1
          maxLength: 50
          example: "Gruppo Università 2025"
    GroupPhoto:
      description: |-
          This schema represents the photo of a group. Hence it is characterized by a single property(the photo itself).
      type: object
      required:
        - photo
      properties:
        photo:
          type: string
          format: binary
          minLength: 1
          maxLength: 100000000
          description: |-
              Group photo file (JPEG, PNG, etc.)
    GroupPhotoUpdated:
      description: |-
          This schema represents the new picture of a particular group(Photo Updated). Hence, it is characterized by the group Id (the id of the group whose pic has been modified), and the new pic (photoUrl).
      type: object
      properties:
        groupId:
          description: |-
              The id of the group.
          type: integer
          example: 42
        photoUrl:
          description: |-
              URL of the uploaded photo
          type: string
          format: url
          example: "https://example.com/photos/group42.jpg"
...
security:
  - bearerAuth: []